requirements.txt
fastapi
uvicorn
opencv-python-headless
pyyaml
numpy
pillow
aiofiles
python-dotenv
